backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
pydantic[email]
loguru
pytest
pytest-asyncio 
httpx
pwdlib[argon2]
pyjwt
python-multipart
pydantic-settings
tzdata